warehouse/requirements.txt
snowflake-connector-python
google-cloud-bigquery
pandas-gbq
pyarrow
pandas
python-dotenv
